ctr_maximizer_gpt/requirements.txt

numpy>=1.24
pandas>=2.0
pyarrow>=12
polars>=0.20
scikit-learn>=1.3
scipy>=1.10
tqdm>=4.66
pyyaml>=6.0

xgboost>=2.0
catboost>=1.2

torch>=2.2

optuna>=3.4
